Create Hovel v.02
v.02 is a new design that follows oo modularity much closer and enables
some of the functionality to be more easily designed. The gameboard is
coming together now with a option panel that comes up when a tile is
right clicked. This is where the player will choose what they want to do
with a particular tile.